Kotlin - The News App
This application makes it easy for users to follow current news and save interesting articles. You can browse the latest news on the home screen and examine it in detail. You can save the news you like and access it quickly later. Thanks to its user-friendly interface and smooth transition animations, navigating the application is quite enjoyable and simple. Aiming to provide a seamless news experience on Android devices, this application is an ideal option for those who want to access information quickly and easily.